Issues/1121

+++ js/ui/appDisplay.js
@@ -1695,31 +1695,24 @@ const AppStoreIcon = new Lang.Class({
         }
     },
 
-    _showDeleteConfirmation: function(draggedSource, deleteCallback) {
-        draggedSource.blockHandler = true;
+    _removeItem: function(source) {
+        source.blockHandler = true;
         this.blockHandler = true;
-        let trashPopup = new TrashPopup({
-            onCancel: Lang.bind(this, function() {
-                this._restoreTrash(trashPopup, draggedSource);
-            }),
-            onAccept: Lang.bind(this, function() {
-                this._restoreTrash(trashPopup, draggedSource);
-                IconGridLayout.layout.removeIcon(draggedSource.getId());
-                if (deleteCallback) {
-                    deleteCallback();
-                }
-            }),
-        });
-        this.actor.set_child(trashPopup.actor);
-    },
 
-    _restoreTrash: function(trashPopup, source) {
-        trashPopup.actor.visible = false;
+        // store the location of the removed item in order to undo it
+        let [ folderId, idx ] = IconGridLayout.layoyt.getPositionForIcon(souce.getId());
+        this._removedItemFolder = folderId;
+        this._removedItemPos = idx;
+
+        IconGridLayout.layout.removeIcon(source.getId());
+
         source.blockHandler = false;
         this.blockHandler = false;
+
         if (source.handleViewDragEnd) {
             source.handleViewDragEnd();
         }
+
         this.handleViewDragEnd();
     },
 
@@ -1733,14 +1726,42 @@ const AppStoreIcon = new Lang.Class({
         return canDelete;
     },
 
+    _deleteItem: function(source) {
+        this._removedItemPos = -1;
+        this._removedItemFolder = null;
+
+        if (source.app) {
+            let appInfo = source.app.get_app_info();
+            if (this._canDelete(appInfo)) {
+                appInfo.delete();
+            }
+        }
+
+        if (source.folder) {
+            if (this._canDelete(source.folder)) {
+                source.folder.delete();
+            }
+        }
+    },
+
+    _undoRemoveItem: function(source) {
+        let pos = this._removedItemPos;
+        let folderId = this._removedItemFolder;
+
+        IconGridLayout.layout.repositionIcon(source.getId(), pos, folderId);
+
+        this._removedItemPos = -1;
+        this._removedItemFolder = null;
+    },
+
     _acceptAppDrop: function(source) {
-        let appInfo = source.app.get_app_info();
-        this._showDeleteConfirmation(source,
-                                     Lang.bind(this, function() {
-                                         if (this._canDelete(appInfo)) {
-                                             appInfo.delete();
-                                         }
-                                     }));
+        this._removeItem(source);
+
+        Main.overview.setMessage(_("%s has been deleted").format(source.app.get_name()),
+                                 { forFeedback: true,
+                                   destroyCallback: Lang.bind(this, this._deleteItem, source),
+                                   undoCallback: Lang.bind(this, this._undoRemoveItem, source)
+                                 });
     },
 
     _acceptFolderDrop: function(source) {
@@ -1759,12 +1780,13 @@ const AppStoreIcon = new Lang.Class({
         }
 
         if (isEmpty) {
-            this._showDeleteConfirmation(source,
-                                         Lang.bind(this, function() {
-                                             if (this._canDelete(folder)) {
-                                                 folder.delete();
-                                             }
-                                         }));
+            this._removeItem(source);
+
+            Main.overview.setMessage(_("%s has been deleted").format(folder.get_name()),
+                                     { forFeedback: true,
+                                       destroyCallback: Lang.bind(this, this._deleteItem, source),
+                                       undoCallback: Lang.bind(this, this._undoRemoveItem, source)
+                                     });
             return;
         }
